pyproject.toml
[project]
name = "xauto"
version = "0.1.0"
description = "X (Twitter) automation utilities"
readme = false
requires-python = ">=3.9"
dependencies = [
  "tweepy>=4.14.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
xauto = "xauto.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["xauto*"]
